feat(modularization): S-4 extract kanban view to web/modules/kanban.js
Extracts ~1440 lines of kanban board logic from app.js into a dedicated ES
module. Shared mutable state is exposed via a kanbanState proxy object with
getter/setter accessors so the ideas page and Gantt IIFE can read live values
without holding their own copies. DnD functions remain in app.js until S-5.

Key design decisions:
- kanbanState getter/setters: single source of truth in the module; ideas
  page and Gantt read kanbanState.cards / .assignees / .breakdownMode etc.
- initKanban() DI: openModal, closeModal, wireColumn, wireCardTouch, and
  loadIdeasPage injected on boot so kanban.js has no implicit globals.
- showBreakdownModal exported for the ideas page openIdeaBreakdown call.
- Gantt hook fixed: window.renderKanban was broken since S-1; replaced with
  window._onKanbanRefresh hook fired at the end of loadKanban().

2868/2868 tests pass. /modules/kanban.js -> 200 application/javascript.
